Move chart area to its own component
