+++ apps/yapms/src/routes/app/imported/+page.svelte
@@ -1,9 +1,4 @@
 <script lang="ts">
-	import CandidateBoxContainer from '$lib/components/candidatebox/CandidateBoxContainer.svelte';
-	import { ChartPositionStore, ChartTypeStore } from '$lib/stores/Chart';
-	import { CandidatesStore } from '$lib/stores/Candidates';
-	import HorizontalBattleChart from '$lib/components/chartbar/battlechart/BattleChart.svelte';
-	import ChartBar from '$lib/components/chartbar/ChartBar.svelte';
 	import { loadRegionsForApp } from '../[country]/[map]/[year]/initialize/LoadRegions';
 	import applyPanZoom from '../[country]/[map]/[year]/initialize/ApplyPanZoom';
 	import { ImportModalStore } from '$lib/stores/Modals';
@@ -22,47 +17,15 @@
 	}
 </script>
 
-<div
-	class="flex flex-grow basis-9/12"
-	class:flex-col-reverse={$ChartPositionStore === 'bottom'}
-	class:flex-row={$ChartPositionStore === 'left'}
->
+{#if $ImportedSVGStore.loaded === false}
+	<h1>Loading Map...</h1>
+{:else}
 	<div
-		class="flex justify-center items-center ml-3 mr-3 mt-3 mb-3"
-		class:hidden={$ChartTypeStore === 'none'}
+		use:setupMap
+		id="map-div"
+		class="overflow-hidden h-full"
+		class:insetsHidden={$MapInsetsStore.hidden}
 	>
-		{#if $ChartTypeStore === 'battle' && $CandidatesStore.length <= 2}
-			<HorizontalBattleChart />
-		{:else if $ChartTypeStore === 'pie'}
-			<ChartBar />
-		{:else}
-			<ChartBar />
-		{/if}
+		{@html $ImportedSVGStore.content}
 	</div>
-
-	<div
-		class="divider"
-		class:divider-vertical={$ChartPositionStore === 'bottom'}
-		class:h-0={$ChartPositionStore === 'bottom'}
-		class:mb-0={$ChartPositionStore === 'bottom'}
-		class:mt-0={$ChartPositionStore === 'bottom'}
-		class:divider-horizontal={$ChartPositionStore === 'left'}
-		class:w-0={$ChartPositionStore === 'left'}
-		class:mr-0={$ChartPositionStore === 'left'}
-		class:ml-0={$ChartPositionStore === 'left'}
-	/>
-
-	<div class="overflow-hidden w-full h-full">
-		<CandidateBoxContainer />
-		<div
-			use:setupMap
-			id="map-div"
-			class="overflow-hidden h-full"
-			class:insetsHidden={$MapInsetsStore.hidden}
-		>
-			{#if $ImportedSVGStore.loaded}
-				{@html $ImportedSVGStore.content}
-			{/if}
-		</div>
-	</div>
-</div>
+{/if}
